Move LoanModel to file
