+++ lib/src/features/loans/data/loans_view_model.dart
@@ -1,10 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
-import 'package:librarian_app/src/features/borrowers/data/borrower_model.dart';
 import 'package:librarian_app/src/features/common/data/lending_api.dart';
 
+import 'loan_model.dart';
 import 'loans_mapper.dart';
-import 'things_model.dart';
 
 class LoansViewModel extends ChangeNotifier {
   LoansViewModel() {
@@ -94,29 +93,4 @@ class LoansViewModel extends ChangeNotifier {
 
     await refresh();
   }
-}
-
-class LoanModel {
-  final String id;
-  final Thing thing;
-  final BorrowerModel borrower;
-  final DateTime checkedOutDate;
-  DateTime dueDate;
-  DateTime? checkedInDate;
-
-  bool get isOverdue {
-    final now = DateTime.now();
-    return DateUtils.dateOnly(dueDate).isBefore(DateUtils.dateOnly(now));
-  }
-
-  bool get isDueToday => DateUtils.isSameDay(DateTime.now(), dueDate);
-
-  LoanModel({
-    required this.id,
-    required this.thing,
-    required this.borrower,
-    required this.checkedOutDate,
-    required this.dueDate,
-    this.checkedInDate,
-  });
 }
