+++ lib/src/features/loans/presentation/loan_details_page.dart
@@ -5,6 +5,7 @@ import 'package:librarian_app/src/features/loans/presentation/loan_details.dart'
 import 'package:provider/provider.dart';
 
 import '../../dashboard/presentation/mobile_layout.dart';
+import '../data/loan_model.dart';
 
 class LoanDetailsPage extends StatefulWidget {
   const LoanDetailsPage(this.loan, {super.key});
